Reject invalid hostnames with low content bytes.
The domain state machines failed to guard their `lablen` transitions
with `not_in_label`.  For the base `domain` machine this was largely
irrelevant, as any byte values are allowed as label content.  For the
`hostname` machine however it caused the machine to incorrectly begin a
new label whenever it encountered an otherwise-invalid byte in the 1-63
range.

This commit corrects the machine to guard the `lablen` transitions and
adds a covering test for the `dns/domain-hostname?` function.